Fixes copy part / copy object encoding issues
Adds integration tests for CopyParts API.
APIs must return "application/xml".
Incoming key reference must undergo decode&encode cycle just like the
destinationFile name we take from the incoming request path. ObjectRef
now does this for every incoming request. Before, clients of ObjectRef
would need to remember to encode the key again before usage. D'OH!
From/To values are of type "long" in the request, not "int". Copying
of parts larger than Integer.MAX_VALUE would have failed.

Special thanks to user @wwelling for spotting these issues and providing
the fix, I just refactored and added an integration test.

fixes #279, makes #278 obsolete.

+++ server/src/main/java/com/adobe/testing/s3mock/domain/FileStore.java
@@ -1120,8 +1120,8 @@ public class FileStore {
    */
   public String copyPart(final String bucket,
       final String key,
-      final int from,
-      final int to,
+      final long from,
+      final long to,
       final String partNumber,
       final String destinationBucket,
       final String destinationFilename,
@@ -1137,10 +1137,10 @@ public class FileStore {
 
   private String copyPart(final String bucket,
       final String key,
-      final int from,
-      final int to,
+      final long from,
+      final long to,
       final File partFile) throws IOException {
-    final int len = to - from + 1;
+    final long len = to - from + 1;
     final S3Object s3Object = resolveS3Object(bucket, key);
 
     try (final InputStream sourceStream = FileUtils.openInputStream(s3Object.getDataFile());
